[cleanup] move dialogs to dialogs folder
