+++ dialogs/aboutdialog.h
