+++ dialogs/plotdialog.cpp
