+++ dialogs/plotdialog.h
