+++ dialogs/plotdialog.ui
