+++ dialogs/tabledialog.ui
